pyproject.toml
[project]
name = "freshpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansible>=12.0.0",
    "ansible-runner>=2.4.1",
    "archinstall>=3.0.11",
    "psutil>=7.1.0",
    "pyudev>=0.24.3",
    "pyyaml>=6.0.3",
]
